requirements.txt

reportlab
arabic-reshaper
python-bidi
pyinstaller